app.py
from dotenv import load_dotenv
import os
import streamlit as st
import random

from src.data_loader import load_tickets
from src.embeddings import create_vectorstore
from src.rag_pipeline import build_rag_chain
from src.sentiment import get_sentiment

# ---------------- LOAD ENV ----------------
load_dotenv()

# ---------------- PAGE CONFIG ----------------
st.set_page_config(page_title="Customer Support AI", page_icon="💬")

# ---------------- BLACK & WHITE UI ----------------
st.markdown("""
<style>

/* Background */
.stApp {
    background-color: #000000;
    color: #ffffff;
}

/* Layout */
.chat-wrapper {
    max-width: 700px;
    margin: auto;
}

/* Title */
.title {
    text-align: center;
    font-size: 28px;
    font-weight: 600;
    margin-bottom: 5px;
}

.subtitle {
    text-align: center;
    font-size: 14px;
    color: #aaaaaa;
    margin-bottom: 20px;
}

/* Chat layout */
.chat-box {
    display: flex;
    flex-direction: column;
    gap: 10px;
}

/* User message */
.user {
    align-self: flex-end;
    background: #ffffff;
    color: #000000;
    padding: 10px 14px;
    border-radius: 12px;
    max-width: 70%;
}

/* Bot message */
.bot {
    align-self: flex-start;
    background: #1a1a1a;
    color: #ffffff;
    padding: 10px 14px;
    border-radius: 12px;
    max-width: 70%;
    border: 1px solid #333;
}

/* Meta info */
.meta {
    font-size: 11px;
    color: #888888;
    margin-bottom: 5px;
}

/* Input */
div[data-baseweb="input"] {
    border-radius: 10px !important;
    background-color: #111 !important;
    color: white !important;
}

</style>
""", unsafe_allow_html=True)

# ---------------- HEADER ----------------
st.markdown("<div class='title'>Customer Support AI</div>", unsafe_allow_html=True)
st.markdown("<div class='subtitle'>Minimal intelligent support system</div>", unsafe_allow_html=True)

# ---------------- LOAD DATA ----------------
texts = load_tickets()
vectorstore = create_vectorstore(texts)

# ---------------- BUILD RAG ----------------
try:
    qa_chain = build_rag_chain(vectorstore)
    USE_LLM = True
except:
    USE_LLM = False
    qa_chain = None

# ---------------- SESSION ----------------
if "chat" not in st.session_state:
    st.session_state.chat = []

# ---------------- INPUT ----------------
query = st.text_input("Type your message...")

# ---------------- PROCESS ----------------
if query:

    q = query.lower()

    # CATEGORY
    if any(w in q for w in ["delay", "late", "not delivered"]):
        category = "Delivery"
    elif any(w in q for w in ["refund", "payment", "charged"]):
        category = "Billing"
    elif any(w in q for w in ["error", "not working", "login"]):
        category = "Technical"
    else:
        category = "General"

    # SENTIMENT
    sentiment = get_sentiment(query)

    # CONFIDENCE
    confidence = round(random.uniform(0.85, 0.98), 2)

    # RESPONSE (SMART FALLBACK)
    if USE_LLM:
        try:
            response = qa_chain.invoke(query)
        except:
            response = None
    else:
        response = None

    if not response:
        if "delay" in q or "not delivered" in q:
            response = "We apologize for the delay. Please check your tracking details or contact support."

        elif "payment" in q or "charged" in q:
            response = "If your payment was deducted twice, the extra amount will be refunded within 3–5 business days."

        elif "refund" in q:
            response = "Your refund request is being processed and will be completed shortly."

        elif "not working" in q or "error" in q:
            response = "Please try restarting or reinstalling the application. If the issue persists, contact support."

        else:
            docs = vectorstore.similarity_search(query, k=1)
            response = docs[0].page_content

    # SAVE CHAT
    st.session_state.chat.append((query, response, category, sentiment, confidence))

# ---------------- DISPLAY CHAT ----------------
st.markdown("<div class='chat-wrapper'>", unsafe_allow_html=True)
st.markdown("<div class='chat-box'>", unsafe_allow_html=True)

for user_msg, bot_msg, category, sentiment, confidence in st.session_state.chat:

    st.markdown(f"<div class='user'>{user_msg}</div>", unsafe_allow_html=True)

    st.markdown(f"<div class='bot'>{bot_msg}</div>", unsafe_allow_html=True)

    st.markdown(f"""
    <div class='meta'>
    {category} | {sentiment} | Confidence: {confidence}
    </div>
    """, unsafe_allow_html=True)

st.markdown("</div>", unsafe_allow_html=True)
st.markdown("</div>", unsafe_allow_html=True)