requirements.txt
streamlit
python-dotenv
langchain-core
langchain-community
langchain-openai
pandas
numpy
scikit-learn